Working usb device

+++ debug.c
@@ -5,7 +5,7 @@ void debug_print_buffer(
     const uint8_t *data,
     const size_t data_len
 ) {
-    for (int i = 0; i < data_len; i++) {
+    for (size_t i = 0; i < data_len; i++) {
         printf("%02x ", data[i]);
     }
     printf("\n");
